+++ search-films/src/components/Paragraph/index.jsx
@@ -1,12 +1,13 @@
 import { default as CssUtils } from '../shared/CssUtils.js';
-import './styles/index.css';
+import styles from './index.module.css';
 
 function Paragraph({text, fontSizeInPx}) {
+	// здесь стили вот так задаются, чтобы на маленьких экранах было норм
 	let style = {
 		fontSize: CssUtils.addPxToCssProperty(fontSizeInPx)
 	};
 	return (
-		<p className="paragraph" style={style}>{text}</p>
+		<p className={styles.paragraph} style={style}>{text}</p>
 	);
 }