+++ search-films/src/components/index.jsx
@@ -1,4 +1,10 @@
-import Logo from '../components/Logo';
-import Menu from '../components/Menu';
-export { Logo, Menu };
+import Flex from './Flex';
+import Input from './Input';
+import LinkedMenuItem from './LinkedMenuItem';
+import Logo from './Logo';
+import Menu from './Menu';
+import MovieCard from './MovieCard';
+import Paragraph from './Paragraph';
+import Title from './Title';
+export { Flex, Input, LinkedMenuItem, Logo, Menu, MovieCard, Paragraph, Title };