+++ search-films/src/layouts/Header/index.jsx
@@ -1,18 +1,20 @@
+import cs from 'classnames';
 import { Logo, Menu } from '../../components';
-import './styles/index.css';
-
+import styles from './index.module.css';
 /**
  * Layout Header component 
  * @param {boolean} isAuthenticated - authenticated user or not
+ * @param {string} userName - (byDefault=null) name of the user
+ * @param {callback} logout - (byDefault=null) can be used to log out
  * @returns {component} Header 
  * 
 */
-function Header({isAuthenticated=false}) {
+function Header() {
 	return (
-		<header className='header'>
-			<div className='header-content restrict-content-size'>
+		<header className={styles.header}>
+			<div className={cs(styles['header-content'], 'restrict-content-size')}>
 				<Logo/>
-				<Menu isAuthenticated={isAuthenticated}/>
+				<Menu/>
 			</div>
 		</header>
 	);