+++ search-films/src/pages/index.jsx
@@ -1,0 +1,2 @@
+export { default as FilmsPage } from './FilmsPage';
+export { default as LoginPage } from './LoginPage';